Assignment4/requirements.txt

numpy==1.19.2
pandas==1.1.3
pytorch==1.8.2
torchvision==0.9.2
tqdm==4.50.2